backend/pyproject.toml
[project]
name = "forum-digest-backend"
version = "0.1.0"
description = "Forum digest platform backend powered by FastAPI and MySQL."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "pymysql>=1.1.1,<2.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "apscheduler>=3.10.4,<4.0.0",
  "httpx>=0.28.0,<1.0.0",
  "feedparser>=6.0.11,<7.0.0",
  "beautifulsoup4>=4.12.3,<5.0.0",
  "langdetect>=1.0.9,<2.0.0",
  "python-dateutil>=2.9.0,<3.0.0",
  "deep-translator>=1.11.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

